server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.6",
    "flake8>=7.3.0",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ruff>=0.12.12",
]

[tool.ruff]
line-length = 79
exclude = ["migrations"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ARG001",
    "BLE001",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "ERA001",
    "DTZ011",
    "PTH100",
    "PTH118",
    "PTH120",
    "RUF005",
    "RUF012",
    "T201",
]
unfixable = ["F401"]
